Rename hook to supports feature
