+++ packages/block-editor/src/hooks/block-renaming.js
