+++ packages/block-editor/src/hooks/index.js
@@ -19,9 +19,9 @@ import './position';
 import './layout';
 import './content-lock-ui';
 import './metadata';
-import './metadata-name';
 import './custom-fields';
 import './block-hooks';
+import './block-renaming';
 import './block-rename-ui';
 
 export { useCustomSides } from './dimensions';
